Fix "nb", examples, references; simplify def lists
Asciidoctor provides the "+" notation to keep any blocks into one single
list item, but markup the lists of definitions would be very difficult,
because of the many possible combinations. At the moment, the simplest
solution is to fake those lists using hardcoded numbers and bold style,
instead of the actual lists markup.